review(#5443): real guard instead of a dead one, shared IT base, rolling-upgrade note
Three points from review.

The page-count fallback read as a safety net that could never fire:
getFileById() THROWS on an unknown id rather than returning null, so the
"component != null ? ... : file.getTotalPages()" branch was unreachable, and a
non-paginated id would have died on the cast instead of degrading. It now asks
getFileByIdIfExists() and tests the type, the same shape snapshotPageCounts()
uses, and the fallback says out loud that the on-disk count may be short - that
number is exactly what made this bug silent, so it must not be taken quietly. No
current caller can reach it: the ids come from addFiles or from
snapshotPageCounts(), which already filtered on PaginatedComponent.

The two ITs were the same 115 lines twice, differing only in one config line, so
a change to the completeness assertions had to be mirrored by hand. The scenario
moves to BaseCompactionIndexCompletenessTest and each IT keeps only what makes it
distinct: its name, its documentation of which cause it covers, and for the split
variant the buffer size that forces the split.

The codec's rolling-upgrade exposure is now written down where the flag is
defined: a node still on the older codec keeps the pre-fix behaviour for split
entries, so the fix only takes effect there once it is upgraded, and the symptom
is silent. Says what to do about a node that was live through a compaction under
the old codec.
